+++ R/goodness.R
@@ -133,6 +133,7 @@ summary.goodness <- function(object, check = "", dec = 2, ...) {
 #' @param x Return value from \code{\link{goodness}}
 #' @param check Show plots for variable var. "observed" for the observed frequencies table, "expected" for the expected frequencies table (i.e., frequencies that would be expected if the null hypothesis holds), "chi_sq" for the contribution to the overall chi-squared statistic for each cell (i.e., (o - e)^2 / e), and "dev_std" for the standardized differences between the observed and expected frequencies (i.e., (o - e) / sqrt(e))
 #' @param shiny Did the function call originate inside a shiny app
+#' @param custom Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This opion can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org/} for options.
 #' @param ... further arguments passed to or from other methods
 #'
 #' @examples
@@ -144,7 +145,7 @@ summary.goodness <- function(object, check = "", dec = 2, ...) {
 #' @seealso \code{\link{summary.goodness}} to summarize results
 #'
 #' @export
-plot.goodness <- function(x, check = "", shiny = FALSE, ...) {
+plot.goodness <- function(x, check = "", shiny = FALSE, custom = FALSE, ...) {
 
 	object <- x; rm(x)
   if (is.character(object)) return(object)
@@ -205,6 +206,9 @@ plot.goodness <- function(x, check = "", shiny = FALSE, ...) {
 		    ylab("")
 	}
 
-	sshhr( do.call(gridExtra::grid.arrange, c(plot_list, list(ncol = 1))) ) %>%
-	  { if (shiny) . else print(.) }
+  if (custom)
+    if (length(plot_list) == 1) return(plot_list[[1]]) else return(plot_list)
+
+	sshhr(gridExtra::grid.arrange(grobs = plot_list, ncol = 1)) %>%
+	  {if (shiny) . else print(.)}
 }